ctype.h (and cctype) must define functions in global namespace for C++.
My mistake... for some reason I thought that <cctype> only
defined the functions in namespace std.

But correct is that, as it was before, in C, ctype.h defines
macro's, and in C++, it defines functions in global namespace.
While cctype in C++, include ctype.h (defining functions in
global namespace) AND includes them in namespace std with
the usual namespace std { using std::isspace; ... etc list.