Workaround for SD write when SD card is ejected
When the SD card has been ejected, a write() to the card never returns.

This only happens if CONFIG_SDIOWAIT_WRCOMPLETE is set.

It turns out that in the blocking case waitevents and wkupevent are both
0, hence we get stuck later waiting for a semaphore.

The "fix" is to check waitevents and wkupevent for 0 and return with an
error for this case.

The other changes are drive-by changes but did no seem to make a
difference.