[BACKPORT] drivers/serial: fix Rx interrupt enable for cdcacm
This is addressing an issue where Rx interrupts were not restored for
cdcacm after the buffer had been filled (in a burst).
The Rx interrupts were only restored after the watchdog timeout of 200ms
fired.

This happened because CONFIG_SERIAL_RXDMA was set, however, for the
cdcacm driver this does not actually matter as it is relying on
uart_enablerxint to be called via the ops table.

This patch makes sure that uart_enablerxint and uart_disablerxint are
always called, independent of CONFIG_SERIAL_RXDMA, relying on the ops
table to be correct for the case where it should be only a no-op.

The issues is that DMA is per port by TX or RX. So all 4
permutations (NO DMA, RX & TX DMA, RX Only DMA or TX Only DMA)
can be enabled on different ports in the low level drivers by port.

So U[S]ART1 may have RX & TX DMA and U[S]ART2 may have NO DMA.
But CONFIG_SERIAL_RXDMA will be lit and then the proper functions
are not called on the ports not using DMA.